refactor: refactor code with react query key and change directory structure
